ResponseEntity

+++ books_crud/src/main/java/com/hastycode/books_crud/controller/BookController.java
@@ -2,9 +2,6 @@ package com.hastycode.books_crud.controller;
 
 import com.hastycode.books_crud.model.Book;
 import com.hastycode.books_crud.service.BookService;
-import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -14,8 +11,11 @@ import java.util.List;
 @RestController
 public class BookController {
 
-    @Autowired
-    private BookService service;
+    private final BookService service;
+
+    public BookController(BookService service) {
+        this.service = service;
+    }
 
     @GetMapping("/home")
     public String home() {
@@ -23,22 +23,56 @@ public class BookController {
     }
 
     @GetMapping("/books")
-    public List<Book> getBooks() {
-        return service.getBooks();
+    public ResponseEntity<List<Book>> getBooks() {
+        return new ResponseEntity<>(service.getBooks(), HttpStatus.OK);
+    }
+
+    @GetMapping("/books/{id}")
+    public ResponseEntity<Book> getBookById(@PathVariable long id) {
+        Book book = null;
+        book = service.getBookById(id);
+
+        if (book != null) {
+            return new ResponseEntity<>(book, HttpStatus.OK);
+        } else {
+            return  new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+        }
     }
 
     @PostMapping("/books")
-    public Book addBook(@RequestBody Book book) {
-        return service.addBook(book);
+    public ResponseEntity<Book> addBook(@RequestBody Book book) {
+        Book newBook = null;
+        newBook = service.addBook(book);
+
+        if (newBook != null) {
+            return new ResponseEntity<>(newBook, HttpStatus.CREATED);
+        } else {
+            return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+        }
+
     }
 
     @PutMapping("/books/{id}")
-    public Book updateBook(@PathVariable long id, @RequestBody Book book) {
-        return service.updateBook(id);
+    public ResponseEntity<Book> updateBook(@PathVariable long id, @RequestBody Book book) {
+        Book newBook = null;
+        newBook = service.updateBook(id, book);
+
+        if(newBook != null) {
+            return new ResponseEntity<>(service.updateBook(id, book), HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+        }
     }
 
     @DeleteMapping("/books/{id}")
-    public void deleteBook(@PathVariable long id) {
-        service.deleteBook(id);
+    public ResponseEntity<String> deleteBook(@PathVariable long id) {
+        Book book = service.getBookById(id);
+
+        if (book != null) {
+            service.deleteBook(id);
+            return new ResponseEntity<>("Deleted Successfully", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("Book not found!", HttpStatus.NOT_FOUND);
+        }
     }
 }
